dom/base/test/jsmodules/importmaps/test_externalImportMap.html

<!DOCTYPE html>
<meta charset=utf-8>
<title>Test an external import map</title>
<script src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>

<!--Import maps spec doesn't clearly define the format of an external import map script.-->
<script src="external_importMap.js" type="importmap" onload="scriptLoaded()" onerror="scriptError()"></script>

<script>
function testLoaded() {
  SimpleTest.waitForExplicitFinish();
  SimpleTest.finish();
}

function scriptLoaded() {
  ok(false, "Loading external import map script should have failed.");
}

function scriptError() {
  ok(true, "Loading external import map script failed.");
}
</script>
<body onload='testLoaded()'></body>
